replace siteTitle and…
replace siteTitle with image tag and add gatsbyLogo as src. in this case, the src does not accept a string or path, but a reference to the import placed at the top of the file. alt tag is set as normal. You can add inline styles here as shown, but this will be temporary as we will move to another way to add CSS to html