create an archive file and…
create an archive file in the components directory. past the content from layout.js into this file to start. remove header and styles because we don’t need the header in this sidebar element, and we will be creating seperate styles for this element.

Change areas containing Layout to Archive including our const and export. Remove renderProps ( they are not needed here ). We can now begin. Detailed notes of what else we did is in our text file 9 Our Query In React