Add links and update our query for order…
We added an unordered list around our list, added the necessary key to allow react to generate an individual key per list item. We then imported Link ( the same way in our index.js file  but added to our existing gatsby import ). In this link we we add our to={} which contains the path containing the directory and the slug.

More detailed notes can be found in text file 10 Improving Our Query with Filtering and Sorting