begin to build our page…
we return to our node file to add a new Promise. we create a new async function that will loop over createPages after running queries. we do this by returning our new Promise.

run graphql and run it with our new query.
the results are coming back with the Promise .then
.then will give us access to the results.
results is an arrow function Promise.
we are running the graphql function that returns a Promise that is the data coming in as results.
the data will come in exactly like what we see in our GraphQL interface.

The forEach allows us to loop over every item in the array.
forEach is a function that will destructure each individual node.

See # 12 Creating Pages from Markdown Files